Integrate Josh's BLE firmware with hub system
- Added NimBLE support for BLE central functionality
- Fixed all format specifier warnings (uint32_t -> %lu)
- Adjusted partition table for larger binary (2MB app)
- Integrated base_ble.c, log_transfer_client.c from Josh
- Updated CMakeLists.txt with BLE source files
- Created comprehensive README with setup/troubleshooting
- Configured for 4MB flash at 115200 baud rate
- Ready for testing with WM-SHEARS device

+++ watermelonHub/main/hub_controller.cpp
@@ -47,8 +47,8 @@ void HubController::processIncomingData() {
     // Check for new data from BLE
     CutEvent* event = bleHandler->getNextEvent();
     if (event != nullptr) {
-        ESP_LOGI(TAG, "Received event #%d: Lat=%.6f, Lon=%.6f, Force=%.2f", 
-                 event->sequence_id, event->latitude, event->longitude, event->force);
+        ESP_LOGI(TAG, "Received event #%lu: Lat=%.6f, Lon=%.6f, Force=%.2f", 
+                 (unsigned long)event->sequence_id, event->latitude, event->longitude, event->force);
         
         // Store in database
         if (dataManager->storeEvent(*event)) {
